requirements.txt
asgiref==3.8.1
boto3==1.40.1
botocore==1.40.1
cffi==1.17.1
cryptography==45.0.6
Django==5.2.3
django-appconf==1.1.0
django-cryptography-5==2.0.3
django-htmx==1.23.2
django-storages==1.14.6
Faker==37.4.0
gunicorn==23.0.0
jmespath==1.0.1
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
s3transfer==0.13.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.9.0
